Bug 1329668 Handle partial initialized when ScriptLoader is canceled. r=baku
UltraBlame original commit: 8bb63d1c87d7bd31631543a41398eddfc9bb1c31

+++ dom/workers/ScriptLoader.cpp
@@ -1579,13 +1579,17 @@ CacheCreator::DeleteCache()
 {
   AssertIsOnMainThread();
 
-  ErrorResult rv;
+  
+  
+  if (!mCacheStorage) {
+    return;
+  }
 
   
   
+  IgnoredErrorResult rv;
   RefPtr<Promise> promise = mCacheStorage->Delete(mCacheName, rv);
   if (NS_WARN_IF(rv.Failed())) {
-    rv.SuppressException();
     return;
   }
